backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
python-multipart

pandas
openpyxl
pdfplumber
python-docx

torch
transformers
accelerate
bitsandbytes
sentencepiece